refactor: move SpinalSimClockDomainTest
